todo auth using laravel breeze

+++ routes/web.php
@@ -22,14 +22,21 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth'])->name('dashboard');
 
-require __DIR__.'/auth.php';
+require __DIR__ . '/auth.php';
 
-//Todo List project--------------------------------------------------------------------------------------------------
-Route::get('/todos', [TodoController::class, 'index']);
-Route::get('/todos/create', [TodoController::class, 'create']);
-Route::post('/todos/create', [TodoController::class, 'store']);
-Route::get('/todos/{id}/edit', [TodoController::class, 'edit']); //dynamic routing e.g id
-Route::patch('/todos/{id}/update', [TodoController::class, 'update']); //patch (with the previous TodoCreateRequest validation)
-Route::put('/todos/{id}/complete', [TodoController::class, 'complete']);
-Route::delete('/todos/{id}/incomplete', [TodoController::class, 'incomplete']);
-Route::delete('/todos/{id}/delete', [TodoController::class, 'delete']);
+//Todo List project------------------------------------------------------------------------------------------------------------------
+Route::middleware('auth')->group(function () {   //apply laravel breeze built-in middleware
+//There are three ways to apply middleware:
+//1. Here as a group
+
+    Route::get('/todos', [TodoController::class, 'index']);
+    Route::get('/todos/create', [TodoController::class, 'create']);
+    Route::post('/todos/create', [TodoController::class, 'store']);
+    Route::get('/todos/{id}/edit', [TodoController::class, 'edit']); //dynamic routing e.g id
+    Route::patch('/todos/{id}/update', [TodoController::class, 'update']); //patch (with the previous TodoCreateRequest validation)
+    Route::put('/todos/{id}/complete', [TodoController::class, 'complete']);
+    Route::delete('/todos/{id}/incomplete', [TodoController::class, 'incomplete']);
+});
+//2.
+Route::delete('/todos/{id}/delete', [TodoController::class, 'delete'])->middleware('auth');
+//3. In TodoControlles as a constructor and passing as: $this->middleware('auth);
